requirements.txt
azure-core==1.7.0
azure-storage-blob==12.3.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
Flask==1.1.2
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.18
oauthlib==3.1.0
pycparser==2.20
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
